backend/requirements.txt
Django>=2.1.2
django-extensions>=2.1.5
djangorestframework>=3.9.1
django-cors-headers>=2.4.0
